vk_bot/requirements.txt
aiofiles==22.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==24.1.0
choicelib==0.1.5
exceptiongroup==1.2.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pydantic==1.10.17
python-dotenv==1.0.1
sniffio==1.3.1
typing_extensions==4.12.2
vbml==1.1.post1
vkbottle==4.3.12
vkbottle-types==5.131.146.16
watchfiles==0.18.1
yarl==1.9.4
requests
